# openapi.json
{
  "components": {
    "schemas": {
...
      "AssumptionsOut": {
        "properties": {
          "note": {
            "title": "Note",
            "type": "string"
          },
          "sensitivity_range_percent": {
            "title": "Sensitivity Range Percent",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "sensitivity_range_percent",
          "note"
        ],
        "title": "AssumptionsOut",
        "type": "object"
      },
...
      "BatchResultOut": {
        "properties": {
          "assumptions": {
            "$ref": "#/components/schemas/AssumptionsOut"
          },
          "batch_digest": {
            "title": "Batch Digest",
            "type": "string"
          },
          "batch_id": {
            "title": "Batch Id",
            "type": "string"
          },
          "fallback_rate": {
            "title": "Fallback Rate",
            "type": "number"
          },
          "fixed_clock": {
            "title": "Fixed Clock",
            "type": "string"
          },
          "generated_at": {
            "title": "Generated At",
            "type": "string"
          },
          "map_coverage_rate": {
            "title": "Map Coverage Rate",
            "type": "number"
          },
          "policies": {
            "items": {
              "$ref": "#/components/schemas/PolicyMetricsOut"
            },
            "title": "Policies",
            "type": "array"
          },
          "real_vs_simulated": {
            "title": "Real Vs Simulated",
            "type": "string"
          },
          "scenario_count": {
            "title": "Scenario Count",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "sensitivity": {
            "items": {
              "$ref": "#/components/schemas/SensitivityOut"
            },
            "title": "Sensitivity",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "batch_id",
          "seed",
          "scenario_count",
          "batch_digest",
          "fixed_clock",
          "assumptions",
          "real_vs_simulated",
          "map_coverage_rate",
          "fallback_rate",
          "policies",
          "sensitivity",
          "generated_at"
        ],
        "title": "BatchResultOut",
        "type": "object"
      },
...
      "DecisionListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DecisionOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "DecisionListOut",
        "type": "object"
      },
      "DecisionOut": {
        "properties": {
          "advisory_cache_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Cache Key"
          },
          "advisory_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Class"
          },
          "advisory_confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Confidence"
          },
          "advisory_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Provider"
          },
          "advisory_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advisory Rationale"
          },
          "allowed_actions": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Actions",
            "type": "array"
          },
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "decision_hash": {
            "title": "Decision Hash",
            "type": "string"
          },
          "decision_id": {
            "title": "Decision Id",
            "type": "string"
          },
          "decision_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Decision Reasons",
            "type": "array"
          },
          "effect_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect Action"
          },
          "effect_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect State"
          },
          "effective_class": {
            "title": "Effective Class",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "gate_checks": {
            "items": {
              "$ref": "#/components/schemas/GateCheckOut"
            },
            "title": "Gate Checks",
            "type": "array"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "known_reason": {
            "title": "Known Reason",
            "type": "boolean"
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "next_eligible_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Eligible At"
          },
          "payment_id": {
            "title": "Payment Id",
            "type": "string"
          },
          "policy_fingerprint": {
            "title": "Policy Fingerprint",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "reason_map_fingerprint": {
            "title": "Reason Map Fingerprint",
            "type": "string"
          },
          "review_required": {
            "title": "Review Required",
            "type": "boolean"
          },
          "selected_action": {
            "title": "Selected Action",
            "type": "string"
          },
          "triage_rationale": {
            "title": "Triage Rationale",
            "type": "string"
          }
        },
        "required": [
          "decision_id",
          "event_id",
          "payment_id",
          "created_at",
          "reason",
          "effective_class",
          "advisory_class",
          "advisory_rationale",
          "advisory_confidence",
          "advisory_provider",
          "advisory_cache_key",
          "known_reason",
          "review_required",
          "triage_rationale",
          "allowed_actions",
          "selected_action",
          "decision_reasons",
          "next_eligible_at",
          "reason_map_fingerprint",
          "policy_version",
          "policy_fingerprint",
          "decision_hash",
          "amount_minor",
          "currency",
          "method",
          "idempotency_key",
          "effect_action",
          "effect_state",
          "gate_checks"
        ],
        "title": "DecisionOut",
        "type": "object"
      },
      "GateCheckOut": {
        "properties": {
          "explanation": {
            "title": "Explanation",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "passed": {
            "title": "Passed",
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "passed",
          "explanation"
        ],
        "title": "GateCheckOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PolicyMetricsOut": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "class_d_violations": {
            "title": "Class D Violations",
            "type": "integer"
          },
          "customer_contacts": {
            "title": "Customer Contacts",
            "type": "integer"
          },
          "decisions_digest": {
            "title": "Decisions Digest",
            "type": "string"
          },
          "policy": {
            "title": "Policy",
            "type": "string"
          },
          "recovered_count": {
            "title": "Recovered Count",
            "type": "integer"
          },
          "recovered_minor_units": {
            "title": "Recovered Minor Units",
            "type": "integer"
          },
          "recovered_minor_units_per_attempt": {
            "title": "Recovered Minor Units Per Attempt",
            "type": "number"
          },
          "recovery_rate": {
            "title": "Recovery Rate",
            "type": "number"
          },
          "scenario_count": {
            "title": "Scenario Count",
            "type": "integer"
          },
          "wasted_attempts": {
            "title": "Wasted Attempts",
            "type": "integer"
          }
        },
        "required": [
          "policy",
          "scenario_count",
          "recovered_count",
          "recovery_rate",
          "recovered_minor_units",
          "attempts",
          "recovered_minor_units_per_attempt",
          "wasted_attempts",
          "customer_contacts",
          "class_d_violations",
          "decisions_digest"
        ],
        "title": "PolicyMetricsOut",
        "type": "object"
      },
...
      "SensitivityOut": {
        "properties": {
          "batch_digest": {
            "title": "Batch Digest",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "policies": {
            "items": {
              "$ref": "#/components/schemas/PolicyMetricsOut"
            },
            "title": "Policies",
            "type": "array"
          },
          "probability_multiplier": {
            "title": "Probability Multiplier",
            "type": "number"
          }
        },
        "required": [
          "label",
          "probability_multiplier",
          "batch_digest",
          "policies"
        ],
        "title": "SensitivityOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deterministic payment-recovery evidence and read-only operator views.",
    "title": "Salvage read-only API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/batches/{batch_id}/results": {
      "get": {
        "operationId": "batch_results_api_v1_batches__batch_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResultOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Batch Results",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/decisions": {
      "get": {
        "operationId": "decisions_api_v1_decisions_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decisions",
        "tags": [
          "console"
        ]
      }
    },
    "/api/v1/decisions/{decision_id}": {
      "get": {
        "operationId": "decision_detail_api_v1_decisions__decision_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "decision_id",
            "required": true,
            "schema": {
              "title": "Decision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decision Detail",
        "tags": [
          "console"
        ]
      }
    },
...
    "/webhooks/razorpay": {
      "post": {
        "operationId": "razorpay_webhook_webhooks_razorpay_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-razorpay-signature",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Razorpay-Signature"
            }
          },
          {
            "in": "header",
            "name": "x-razorpay-event-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Razorpay-Event-Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "boolean"
                      }
                    ]
                  },
                  "title": "Response Razorpay Webhook Webhooks Razorpay Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Razorpay Webhook",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  }
}